+++ flex-objects.php
@@ -70,9 +70,6 @@ class FlexObjectsPlugin extends Plugin
             'onFormRegisterTypes' => [
                 ['onFormRegisterTypes', 0]
             ],
-            'onAdminRegisterPermissions' => [
-                ['onAdminRegisterPermissions', -10]
-            ],
             RegisterPermissionsEvent::class => [
                 ['onRegisterPermissions', 100]
             ],
@@ -177,29 +174,6 @@ class FlexObjectsPlugin extends Plugin
 
         $permissions = $event->permissions;
         $permissions->addActions($actions);
-    }
-
-    /**
-     * Initial stab at registering permissions (WIP)
-     *
-     * @param Event $e
-     */
-    public function onAdminRegisterPermissions(Event $e)
-    {
-        $admin = $e['admin'];
-        if (!$admin) {
-            return;
-        }
-
-        $permissions = [
-            'admin.flex-object.list' => 'boolean',
-            'admin.flex-object.create' => 'boolean',
-            'admin.flex-object.read' => 'boolean',
-            'admin.flex-object.update' => 'boolean',
-            'admin.flex-object.delete' => 'boolean',
-        ];
-
-        $admin->addPermissions($permissions);
     }
 
     public function onFormRegisterTypes(Event $event): void